requirements.txt
annotated-doc
annotated-types
anyio
fastapi
idna
numpy
pandas
pydantic
pydantic-core
python-dateutil
python-dotenv
requests
scikit-learn==1.6.1
six
starlette
streamlit
typing-extensions
typing-inspection
uvicorn
uv